Speed up stateindex and implement solvers for solving for MDP and QMDP solutions fastly

+++ src/actions.jl
@@ -12,7 +12,7 @@ const ACTION_DIRS = (RSPos(0,1),
                     RSPos(0,0))
 
 POMDPs.actions(pomdp::RockSamplePOMDP{K}) where K = 1:N_BASIC_ACTIONS+K
-POMDPs.actionindex(pomdp::RockSamplePOMDP, a::Int64) = a
+POMDPs.actionindex(pomdp::RockSamplePOMDP, a::Int) = a
 
 function POMDPs.actions(pomdp::RockSamplePOMDP{K}, s::RSState) where K
     if in(s.pos, pomdp.rocks_positions) # slow? pomdp.rock_pos is a vec
